updated sec scan and trivy
