+++ go/main.go
@@ -9,9 +9,12 @@ package main
 import (
 	"context"
 	"dagger/go/internal/dagger"
+	"dagger/go/security"
+	"dagger/go/stats"
 	"encoding/json"
 	"fmt"
 	"strings"
+
 	"time"
 )
 
@@ -21,36 +24,7 @@ type Go struct {
 	KoContainer     *dagger.Container
 }
 
-type WorkflowResult struct {
-	LintOutput  string            // Output from the lint step
-	BuildOutput *dagger.Directory // Output from the build step
-}
-
-type WorkflowStats struct {
-	Lint struct {
-		Duration string   `json:"duration"`
-		Findings []string `json:"findings"`
-	} `json:"lint"`
-	Build struct {
-		Duration   string `json:"duration"`
-		BinarySize string `json:"binarySize"`
-	} `json:"build"`
-	Test struct {
-		Duration string `json:"duration"`
-		Coverage string `json:"coverage"`
-	} `json:"test"`
-	SecurityScan struct {
-		Duration string   `json:"duration"`
-		Findings []string `json:"findings"`
-	} `json:"securityScan"`
-	TrivyScan struct {
-		Duration string   `json:"duration"`
-		Findings []string `json:"findings"`
-	} `json:"trivyScan"`
-	TotalDuration string `json:"totalDuration"` // Total duration of the workflow
-}
-
-// GetGoLangContainer return the default image for golang
+// GetGoLangContainer returns the default image for golang
 func (m *Go) GetGoLangContainer(goVersion string) *dagger.Container {
 	return dag.Container().
 		From("golang:" + goVersion)
@@ -96,6 +70,182 @@ func New(
 	golang.Src = src
 
 	return golang
+}
+
+func (m *Go) RunWorkflowEntryStage(
+	ctx context.Context,
+	src *dagger.Directory,
+	// +optional
+	// +default="500s"
+	lintTimeout string,
+	// +optional
+	// +default="1.23.6"
+	goVersion string,
+	// +optional
+	// +default="linux"
+	os string,
+	// +optional
+	// +default="amd64"
+	arch string,
+	// +optional
+	// +default="main.go"
+	goMainFile string,
+	// +optional
+	// +default="main"
+	binName string,
+	// +optional
+	// +default="2.22.1"
+	secureGoVersion string,
+	// +optional
+	// +default="false"
+	lintCanFail bool, // If true, linting can fail without stopping the workflow
+	// +optional
+	// +default="./..."
+	testArg string, // Arguments for `go test`
+	// +optional
+	// +default="false"
+	securityScanCanFail bool, // If true, security scan can fail without stopping the workflow
+	// +optional
+	// +default="false"
+	trivyScanCanFail bool, // If true, Trivy scan can fail without stopping the workflow
+	// +optional
+	// +default="HIGH,CRITICAL"
+	trivySeverity string, // Severity levels to include (e.g., "HIGH,CRITICAL")
+	// +optional
+	// +default="0.59.1"
+	trivyVersion string,
+) (*dagger.File, error) {
+	// Create a struct to hold the statistics
+	stats := stats.WorkflowStats{}
+
+	// Start timing the workflow
+	startTime := time.Now()
+
+	// Create a channel to collect errors from goroutines
+	errChan := make(chan error, 5) // Buffer size of 5 for lint, build, test, security scan, and Trivy scan
+
+	// Run Lint step in a goroutine
+	go func() {
+		lintStart := time.Now()
+		lintOutput, err := m.Lint(ctx, src, lintTimeout).Stdout(ctx)
+		if err != nil {
+			if !lintCanFail {
+				errChan <- fmt.Errorf("error running lint: %w", err)
+				return
+			}
+			// If lintCanFail is true, log the error but continue
+			stats.Lint.Findings = []string{fmt.Sprintf("Linting failed: %v", err)}
+		} else {
+			stats.Lint.Findings = strings.Split(lintOutput, "\n") // Split lint output into findings
+		}
+		stats.Lint.Duration = time.Since(lintStart).String()
+		errChan <- nil
+	}()
+
+	// Run Build step in a goroutine
+	go func() {
+		buildStart := time.Now()
+		buildOutput := m.Build(ctx, goVersion, os, arch, goMainFile, binName, src)
+		stats.Build.Duration = time.Since(buildStart).String()
+
+		// Calculate binary size
+		binaryPath := binName
+		binarySize, err := buildOutput.File(binaryPath).Size(ctx)
+		if err != nil {
+			errChan <- fmt.Errorf("error getting binary size: %w", err)
+			return
+		}
+		stats.Build.BinarySize = fmt.Sprintf("%d bytes", binarySize)
+		errChan <- nil
+	}()
+
+	// Run Test step in a goroutine
+	go func() {
+		testStart := time.Now()
+		testOutput, err := m.Test(ctx, src, goVersion, testArg)
+		if err != nil {
+			errChan <- fmt.Errorf("error running tests: %w", err)
+			return
+		}
+		stats.Test.Duration = time.Since(testStart).String()
+
+		// Extract coverage from test output
+		coverage := security.ExtractCoverage(testOutput)
+		stats.Test.Coverage = coverage
+		errChan <- nil
+	}()
+
+	// RUN SECURITY SCAN STEP IN A GOROUTINE
+	go func() {
+		securityScanStart := time.Now()
+		reportFile, err := m.SecurityScan(ctx, src, secureGoVersion)
+		if err != nil {
+			if !securityScanCanFail {
+				errChan <- fmt.Errorf("error running security scan: %w", err)
+				return
+			}
+			// If securityScanCanFail is true, log the error but continue
+			stats.SecurityScan.Findings = []string{fmt.Sprintf("Security scan failed: %v", err)}
+		} else {
+			// Read the report file contents
+			reportContent, err := reportFile.Contents(ctx)
+			if err != nil {
+				errChan <- fmt.Errorf("error reading security report: %w", err)
+				return
+			}
+			stats.SecurityScan.Findings = strings.Split(reportContent, "\n") // Split report content into findings
+		}
+		stats.SecurityScan.Duration = time.Since(securityScanStart).String()
+		errChan <- nil
+	}()
+
+	// RUN TRIVY SCAN STEP IN A GOROUTINE
+	go func() {
+		trivyScanStart := time.Now()
+		reportFile, err := m.TrivyScan(ctx, src, trivySeverity, trivyVersion)
+		if err != nil {
+			if !trivyScanCanFail {
+				errChan <- fmt.Errorf("error running Trivy scan: %w", err)
+				return
+			}
+			// If trivyScanCanFail is true, log the error but continue
+			stats.TrivyScan.Findings = []string{fmt.Sprintf("Trivy scan failed: %v", err)}
+		} else {
+			// Read the report file contents
+			reportContent, err := reportFile.Contents(ctx)
+			if err != nil {
+				errChan <- fmt.Errorf("error reading Trivy report: %w", err)
+				return
+			}
+			stats.TrivyScan.Findings = strings.Split(reportContent, "\n") // Split report content into findings
+		}
+		stats.TrivyScan.Duration = time.Since(trivyScanStart).String()
+		errChan <- nil
+	}()
+
+	// Wait for all goroutines to complete
+	for i := 0; i < 5; i++ {
+		if err := <-errChan; err != nil {
+			return nil, err
+		}
+	}
+
+	// Track total workflow duration
+	stats.TotalDuration = time.Since(startTime).String()
+
+	// Generate JSON file with statistics
+	statsJSON, err := json.MarshalIndent(stats, "", "  ")
+	if err != nil {
+		return nil, fmt.Errorf("error generating stats JSON: %w", err)
+	}
+
+	// Write JSON to a file in the container
+	statsFile := dag.Directory().
+		WithNewFile("workflow-stats.json", string(statsJSON)).
+		File("workflow-stats.json")
+
+	// Return the stats file
+	return statsFile, nil
 }
 
 // Lint runs the linter on the provided source code
@@ -194,7 +344,7 @@ func (m *Go) SearchVulnerabilities(
 	}
 
 	// Step 2: Parse the JSON report
-	var report TrivyReport
+	var report security.TrivyReport
 	if err := json.Unmarshal([]byte(scanJSON), &report); err != nil {
 		return nil, fmt.Errorf("error parsing scan report: %w", err)
 	}
@@ -214,183 +364,6 @@ func (m *Go) SearchVulnerabilities(
 	}
 
 	return vulnerabilities, nil
-}
-
-// TrivyReport represents the structure of a Trivy JSON report
-type TrivyReport struct {
-	Results []struct {
-		Vulnerabilities []struct {
-			VulnerabilityID string `json:"VulnerabilityID"`
-			Severity        string `json:"Severity"`
-		} `json:"Vulnerabilities"`
-	} `json:"Results"`
-}
-
-func (m *Go) RunWorkflowEntryStage(
-	ctx context.Context,
-	src *dagger.Directory,
-	// +optional
-	// +default="500s"
-	lintTimeout string,
-	// +optional
-	// +default="1.23.6"
-	goVersion string,
-	// +optional
-	// +default="linux"
-	os string,
-	// +optional
-	// +default="amd64"
-	arch string,
-	// +optional
-	// +default="main.go"
-	goMainFile string,
-	// +optional
-	// +default="main"
-	binName string,
-	// +optional
-	// +default="false"
-	lintCanFail bool, // If true, linting can fail without stopping the workflow
-	// +optional
-	// +default="./..."
-	testArg string, // Arguments for `go test`
-	// +optional
-	// +default="false"
-	securityScanCanFail bool, // If true, security scan can fail without stopping the workflow
-	// +optional
-	// +default="false"
-	trivyScanCanFail bool, // If true, Trivy scan can fail without stopping the workflow
-) (*dagger.File, error) {
-	// Create a struct to hold the statistics
-	stats := WorkflowStats{}
-
-	// Start timing the workflow
-	startTime := time.Now()
-
-	// Create a channel to collect errors from goroutines
-	errChan := make(chan error, 5) // Buffer size of 5 for lint, build, test, security scan, and Trivy scan
-
-	// Run Lint step in a goroutine
-	go func() {
-		lintStart := time.Now()
-		lintOutput, err := m.Lint(ctx, src, lintTimeout).Stdout(ctx)
-		if err != nil {
-			if !lintCanFail {
-				errChan <- fmt.Errorf("error running lint: %w", err)
-				return
-			}
-			// If lintCanFail is true, log the error but continue
-			stats.Lint.Findings = []string{fmt.Sprintf("Linting failed: %v", err)}
-		} else {
-			stats.Lint.Findings = strings.Split(lintOutput, "\n") // Split lint output into findings
-		}
-		stats.Lint.Duration = time.Since(lintStart).String()
-		errChan <- nil
-	}()
-
-	// Run Build step in a goroutine
-	go func() {
-		buildStart := time.Now()
-		buildOutput := m.Build(ctx, goVersion, os, arch, goMainFile, binName, src)
-		stats.Build.Duration = time.Since(buildStart).String()
-
-		// Calculate binary size
-		binaryPath := binName
-		binarySize, err := buildOutput.File(binaryPath).Size(ctx)
-		if err != nil {
-			errChan <- fmt.Errorf("error getting binary size: %w", err)
-			return
-		}
-		stats.Build.BinarySize = fmt.Sprintf("%d bytes", binarySize)
-		errChan <- nil
-	}()
-
-	// Run Test step in a goroutine
-	go func() {
-		testStart := time.Now()
-		testOutput, err := m.Test(ctx, src, goVersion, testArg)
-		if err != nil {
-			errChan <- fmt.Errorf("error running tests: %w", err)
-			return
-		}
-		stats.Test.Duration = time.Since(testStart).String()
-
-		// Extract coverage from test output
-		coverage := extractCoverage(testOutput)
-		stats.Test.Coverage = coverage
-		errChan <- nil
-	}()
-
-	// Run Security Scan step in a goroutine
-	go func() {
-		securityScanStart := time.Now()
-		securityScanOutput, err := m.SecurityScan(ctx, src)
-		if err != nil {
-			if !securityScanCanFail {
-				errChan <- fmt.Errorf("error running security scan: %w", err)
-				return
-			}
-			// If securityScanCanFail is true, log the error but continue
-			stats.SecurityScan.Findings = []string{fmt.Sprintf("Security scan failed: %v", err)}
-		} else {
-			stats.SecurityScan.Findings = strings.Split(securityScanOutput, "\n") // Split security scan output into findings
-		}
-		stats.SecurityScan.Duration = time.Since(securityScanStart).String()
-		errChan <- nil
-	}()
-
-	// Run Trivy Scan step in a goroutine
-	go func() {
-		trivyScanStart := time.Now()
-		trivyScanOutput, err := m.TrivyScan(ctx, src)
-		if err != nil {
-			if !trivyScanCanFail {
-				errChan <- fmt.Errorf("error running Trivy scan: %w", err)
-				return
-			}
-			// If trivyScanCanFail is true, log the error but continue
-			stats.TrivyScan.Findings = []string{fmt.Sprintf("Trivy scan failed: %v", err)}
-		} else {
-			stats.TrivyScan.Findings = strings.Split(trivyScanOutput, "\n") // Split Trivy scan output into findings
-		}
-		stats.TrivyScan.Duration = time.Since(trivyScanStart).String()
-		errChan <- nil
-	}()
-
-	// Wait for all goroutines to complete
-	for i := 0; i < 5; i++ {
-		if err := <-errChan; err != nil {
-			return nil, err
-		}
-	}
-
-	// Track total workflow duration
-	stats.TotalDuration = time.Since(startTime).String()
-
-	// Generate JSON file with statistics
-	statsJSON, err := json.MarshalIndent(stats, "", "  ")
-	if err != nil {
-		return nil, fmt.Errorf("error generating stats JSON: %w", err)
-	}
-
-	// Write JSON to a file in the container
-	statsFile := dag.Directory().
-		WithNewFile("workflow-stats.json", string(statsJSON)).
-		File("workflow-stats.json")
-
-	// Return the stats file
-	return statsFile, nil
-}
-
-// Helper function to extract coverage from test output
-func extractCoverage(testOutput string) string {
-	// Look for a line like "coverage: 75.0% of statements"
-	lines := strings.Split(testOutput, "\n")
-	for _, line := range lines {
-		if strings.Contains(line, "coverage:") {
-			return strings.TrimSpace(line)
-		}
-	}
-	return "coverage: unknown"
 }
 
 // Returns lines that match a pattern in the files of the provided Directory
@@ -414,9 +387,6 @@ func (m *Go) Build(
 	src *dagger.Directory) *dagger.Directory {
 
 	// MOUNT CLONED REPOSITORY INTO `GOLANG` IMAGE
-
-	PrintDirectoryInfo(ctx, src)
-
 	golang := m.
 		GetGoLangContainer(goVersion).
 		WithDirectory("/src", src).
@@ -440,29 +410,6 @@ func (m *Go) Build(
 	outputDir := golang.Directory(path)
 
 	return outputDir
-}
-
-func PrintDirectoryInfo(ctx context.Context, src *dagger.Directory) {
-	if src == nil {
-		fmt.Println("Directory is nil")
-		return
-	}
-
-	id, err := src.ID(ctx)
-	if err != nil {
-		fmt.Println("Error getting directory ID:", err)
-		return
-	}
-	fmt.Println("Dagger Directory ID:", id)
-
-	// List files inside the directory
-	entries, err := src.Entries(ctx)
-	if err != nil {
-		fmt.Println("Error retrieving directory entries:", err)
-		return
-	}
-
-	fmt.Println("Directory contains:", entries)
 }
 
 // Returns lines that match a pattern in the files of the provided Directory
@@ -535,41 +482,52 @@ func (m *Go) Test(
 func (m *Go) TrivyScan(
 	ctx context.Context,
 	src *dagger.Directory,
-) (string, error) {
+	// +optional
+	// +default="HIGH,CRITICAL"
+	severity string, // Severity levels to include (e.g., "HIGH,CRITICAL")
+	// +optional
+	// +default="0.59.1"
+	trivyVersion string,
+) (*dagger.File, error) {
 	// Create a container with Trivy installed
 	container := dag.Container().
-		From("aquasec/trivy:latest"). // Use the official Trivy image
+		From("aquasec/trivy:"+trivyVersion). // Use the official Trivy image
 		WithDirectory("/src", src).
 		WithWorkdir("/src")
 
-	// Run Trivy to scan the source folder
-	output, err := container.
-		WithExec([]string{"trivy", "fs", "--severity", "HIGH,CRITICAL", "/src"}).
-		Stdout(ctx)
-	if err != nil {
-		return "", fmt.Errorf("error running Trivy: %w", err)
-	}
+	// Run Trivy to scan the source folder and write the output to a file
+	reportPath := "/src/trivy-report.txt"
+	container = container.
+		WithExec([]string{"sh", "-c", fmt.Sprintf("trivy fs --severity %s /src > %s || true", severity, reportPath)}) // Ignore exit code
 
-	return output, nil
+	// Get the Trivy report file
+	reportFile := container.File(reportPath)
+
+	// Return the Trivy report file
+	return reportFile, nil
 }
 
 func (m *Go) SecurityScan(
 	ctx context.Context,
 	src *dagger.Directory,
-) (string, error) {
+	// +optional
+	// +default="2.22.1"
+	secureGoVersion string,
+) (*dagger.File, error) {
 	// Create a container with gosec installed
 	container := dag.Container().
-		From("securego/gosec:latest"). // Use the official gosec image
+		From("securego/gosec:"+secureGoVersion). // Use the official gosec image
 		WithDirectory("/src", src).
 		WithWorkdir("/src")
 
-	// Run gosec to scan the source code
-	output, err := container.
-		WithExec([]string{"gosec", "./..."}).
-		Stdout(ctx)
-	if err != nil {
-		return "", fmt.Errorf("error running gosec: %w", err)
-	}
+	// Run gosec to scan the source code and write the output to a file
+	reportPath := "/src/security-report.txt"
+	container = container.
+		WithExec([]string{"sh", "-c", "gosec ./... > " + reportPath + " || true"}) // Ignore exit code
 
-	return output, nil
+	// Get the security report file
+	reportFile := container.File(reportPath)
+
+	// Return the security report file
+	return reportFile, nil
 }
